+++ CHANGELOG.md
@@ -8,9 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/); this proj
 
 ### Added
 - Added `Play`/`Wait` extension methods for `Animation` and `Animator`.
-- Added `AsyncCreationOptions.UseSharedSynchronizationContext` creation option that allows using a shared synchronization context for the operation continuations.
-- Added `AsyncResult.DefaultCreationOptions` static property to set options used by all operations.
-- Added `AsyncResult.SharedSynchronizationContext` static property to be used by continuations if `AsyncCreationOptions.UseSharedSynchronizationContext` option is set.
+- Added `AsyncResult.SharedSynchronizationContext` static property.
 
 ### Changed
 - Changed `AsyncResult` constructors argument order to avoid ambiguity in some cases.